requirements/base.txt
cloudinary==1.25.0
dj-database-url==0.5.0
dj-static==0.0.6
django==3.2.4
django-allauth==0.44.0
django-crispy-forms==1.12.0
django-localflavor==3.1
django-mathfilters==1.0.0
django-widget-tweaks==1.4.8
fontawesome-free==5.15.3
gunicorn==20.0.4
mercadopago==2.0.7
pytest==6.2.4
requests==2.25.1
slack-sdk==3.7.0
urllib3==1.26.5